requirements.txt
Flask
yagmail
gunicorn
python-dotenv
flask-cors
